splite on multiple files
