+++ README.md
@@ -1,5 +1,7 @@
 # mini_messenger
 ## Описание и инструкция по пользованию
-mini_messenger представляет из себя своего рода Discord (discord lite если хотите). После прохождение простой регистрации или входа в аккаунт(если он у вас уже есть) будет доступна возможность создать новую комнату и посмотреть список уже существующих комнат. Комната представляет собой чат в которой может писать каждый зарегистрировавшийся. Таким образом в новой комнате вы можете общаться с более чем одним человеком.
+**mini_messenger** представляет из себя своего рода *Discord* (*discord-lite* если хотите). После прохождение простой регистрации или входа в аккаунт(если он у вас уже есть) будет доступна возможность создать новую комнату и посмотреть список уже существующих комнат. Комната представляет собой чат в которой может писать каждый зарегистрировавшийся. Таким образом в новой комнате вы можете общаться с более чем одним человеком.
+## Замечания
+**mini_messenger** работает локально, поэтому для его тестирования необходимо запустить **_localhost_** в разных браузерах(или перейти по ссылке в разных браузерах) и войти (если уже есть аккаунты) или зарегистрироваться от разных аккаунтов. Вначале база данных *db.sqlite3* пуста, поэтому необходимо зарегистрировать несколько аккаунтов.
